Basic parser structure
This attempts to parse the input gcode into a sort of tree of
relocatable blocks.  The UnifiedGCode struct contains start and end
blocks, and a slice of layers.  Each layer is a slice of blocks, which
are intermixed model and purge blocks.

From here, it should be straightforward to improve command parsing to
extract movement information (particularly extrude length), and write
block.TotalExtrudeLength().

Next step is to implement some global state for things like
absolute/relative movement and extrusion, selected tool, etc, and have
the command parser update that state on the fly.  For instance, when
parsing a command that contains an E move, the parser could use the
state of M82/83 and the state of the current E value to calculate and
store the E move for that command.