Implement statement parser
The statement parser is simple, it splits a line into code and comment
It then uses two regular expressions to pull out the command and all
parameters, then stores those parameters in a map.

I also implemented the Group type, which classifies g-code statements by
NIST-standard groups (such as MOTION).

I implemented the Statement object with a couple of example methods:
Group() returns the statement's Group, IsMotion() is a boolean check of
Group().  Moved(axis) returns the value of that axis parameter if the
statement is a motion command.

I implemented two test cases which exercise parsing, summing E moves
across a list of statements, and verifying group classification.